+++ src/main/java/com/besson/arknights/ArknightsFurnitureClient.java
@@ -4,10 +4,13 @@ import com.besson.arknights.block.ModBlockEntities;
 import com.besson.arknights.block.ModBlocks;
 import com.besson.arknights.block.custom.renderer.LargeShelfRenderer;
 import com.besson.arknights.block.custom.renderer.SeatRenderer;
+import com.besson.arknights.screen.ModScreenHandlers;
+import com.besson.arknights.screen.SmallSlotScreen;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.blockrenderlayer.v1.BlockRenderLayerMap;
 import net.fabricmc.fabric.api.client.rendering.v1.BlockEntityRendererRegistry;
 import net.fabricmc.fabric.api.client.rendering.v1.EntityRendererRegistry;
+import net.minecraft.client.gui.screen.ingame.HandledScreens;
 import net.minecraft.client.render.RenderLayer;
 import net.minecraft.client.render.block.entity.BlockEntityRendererFactories;
 
@@ -21,6 +24,8 @@ public class ArknightsFurnitureClient implements ClientModInitializer {
         BlockRenderLayerMap.INSTANCE.putBlock(ModBlocks.TRANSPARENT_DIVIDING_PARTITION, RenderLayer.getTranslucent());
 
         EntityRendererRegistry.register(ModBlockEntities.SEAT, SeatRenderer::new);
-        BlockEntityRendererFactories.register(ModBlockEntities.LARGE_SHELF, LargeShelfRenderer::new);
+        BlockEntityRendererFactories.register(ModBlockEntities.LARGE_SHELF_EMPTY, LargeShelfRenderer::new);
+
+        HandledScreens.register(ModScreenHandlers.SMALL_SCREEN_HANDLER, SmallSlotScreen::new);
     }
 }