+++ src/main/java/com/besson/arknights/screen/SmallSlotScreen.java
@@ -1,0 +1,35 @@
+package com.besson.arknights.screen;
+
+import com.besson.arknights.ArknightsFurniture;
+import com.mojang.blaze3d.systems.RenderSystem;
+import net.minecraft.client.gui.DrawContext;
+import net.minecraft.client.gui.screen.ingame.HandledScreen;
+import net.minecraft.client.render.GameRenderer;
+import net.minecraft.entity.player.PlayerInventory;
+import net.minecraft.text.Text;
+import net.minecraft.util.Identifier;
+
+public class SmallSlotScreen extends HandledScreen<SmallSlotScreenHandler> {
+    public static final Identifier TEXTURE = new Identifier(ArknightsFurniture.MOD_ID, "textures/gui/small_slot.png");
+    public SmallSlotScreen(SmallSlotScreenHandler handler, PlayerInventory inventory, Text title) {
+        super(handler, inventory, title);
+    }
+
+    @Override
+    protected void drawBackground(DrawContext context, float delta, int mouseX, int mouseY) {
+        RenderSystem.setShader(GameRenderer::getPositionTexProgram);
+        RenderSystem.setShaderColor(1f,1f,1f,1f);
+        RenderSystem.setShaderTexture(0,TEXTURE);
+        int x = (width - backgroundWidth) / 2;
+        int y = (height - backgroundHeight) / 2;
+
+        context.drawTexture(TEXTURE,x,y,0,0,backgroundWidth,backgroundHeight);
+    }
+
+    @Override
+    public void render(DrawContext context, int mouseX, int mouseY, float delta) {
+        renderBackground(context);
+        super.render(context, mouseX, mouseY, delta);
+        drawMouseoverTooltip(context,mouseX,mouseY);
+    }
+}