+++ src/main/java/com/besson/arknights/screen/SmallSlotScreenHandler.java
@@ -1,0 +1,87 @@
+package com.besson.arknights.screen;
+
+import net.minecraft.block.entity.BlockEntity;
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.entity.player.PlayerInventory;
+import net.minecraft.inventory.Inventory;
+import net.minecraft.item.ItemStack;
+import net.minecraft.network.PacketByteBuf;
+import net.minecraft.screen.ScreenHandler;
+import net.minecraft.screen.slot.Slot;
+
+public class SmallSlotScreenHandler extends ScreenHandler {
+    private final Inventory inventory;
+    public SmallSlotScreenHandler(int syncId, PlayerInventory playerInventory, PacketByteBuf packetByteBuf) {
+        this(syncId, playerInventory, playerInventory.player.getWorld().getBlockEntity(packetByteBuf.readBlockPos()));
+    }
+
+    public SmallSlotScreenHandler(int syncId, PlayerInventory playerInventory, BlockEntity blockEntity){
+        super(ModScreenHandlers.SMALL_SCREEN_HANDLER, syncId);
+        checkSize((Inventory) blockEntity, 3);
+        this.inventory = (Inventory) blockEntity;
+        inventory.onOpen(playerInventory.player);
+
+        for (int i = 0; i < 3; i++){
+            this.addSlot(new Slot(inventory, i, 62 + i * 18, 36));
+        }
+
+        addPlayerInventory(playerInventory);
+        addPlayerHotbar(playerInventory);
+    }
+
+    private void addPlayerInventory(PlayerInventory playerInventory){
+        for (int i = 0; i < 3; ++i) {
+            for (int l = 0; l < 9; ++l) {
+                this.addSlot(new Slot(playerInventory, l + i * 9 + 9, 8 + l * 18, 84 + i * 18));
+            }
+        }
+    }
+
+    private void addPlayerHotbar(PlayerInventory playerInventory){
+        for (int i = 0; i < 9; ++i) {
+            this.addSlot(new Slot(playerInventory, i, 8 + i * 18, 142));
+        }
+    }
+    @Override
+    public ItemStack quickMove(PlayerEntity player, int slot) {
+        ItemStack itemStack = ItemStack.EMPTY;
+        Slot slot2 = this.slots.get(slot);
+        if (slot2 != null && slot2.hasStack()) {
+            ItemStack itemStack2 = slot2.getStack();
+            itemStack = itemStack2.copy();
+            if (slot < 3) {
+                if (!this.insertItem(itemStack2, 3, this.slots.size(), true)) {
+                    return ItemStack.EMPTY;
+                }
+            } else if (!this.insertItem(itemStack2, 0, 3, false)) {
+                return ItemStack.EMPTY;
+            }
+
+            if (itemStack2.isEmpty()) {
+                slot2.setStack(ItemStack.EMPTY);
+            } else {
+                slot2.markDirty();
+            }
+        }
+
+        return itemStack;
+    }
+
+    @Override
+    public boolean canUse(PlayerEntity player) {
+        return this.inventory.canPlayerUse(player);
+    }
+
+    @Override
+    public void onClosed(PlayerEntity player) {
+        super.onClosed(player);
+        this.inventory.onClose(player);
+    }
+
+    public Inventory getInventory() {
+        return this.inventory;
+    }
+    public int getRows() {
+        return 1;
+    }
+}